LPS-49135 Move EhcacheConfigurationalUtil to com.liferay.portal.dao.orm.hibernate.region, since it is only used by hibernate, no changes.
