+++ portal-impl/test/unit/com/liferay/portal/dao/orm/hibernate/region/EhcacheConfigurationUtilTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.ehcache;
+package com.liferay.portal.dao.orm.hibernate.region;
 
 import com.liferay.portal.cache.cluster.HibernatePortalCacheClusterReplicatorFactory;
 import com.liferay.portal.kernel.test.CodeCoverageAssertor;
