+++ portal-impl/test/unit/com/liferay/portal/dao/orm/hibernate/region/dependencies/test-multi-ehcache-config.xml
