+++ portal-impl/test/unit/com/liferay/portal/dao/orm/hibernate/region/dependencies/test-single-ehcache-config.xml
